+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### Fixes
 
-- Attempt to fix NPE with config.
+- More 1.19 related config fixes.
 
 Closed issues: None.
 
-[Full Changelog](https://github.com/JamCoreModding/JamLib/compare/0.0.6...0.0.7)
+[Full Changelog](https://github.com/JamCoreModding/JamLib/compare/0.0.7...0.0.8)